refactor - view package와 domain package 생성 및 클래스 이동
