+++ src/main/java/racingcar/domain/GameNumbers.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class GameNumbers {
     private int tryCounts;
