+++ src/test/java/racingcar/CarControllerTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarController;
+import racingcar.domain.CarList;
 
 import java.util.ArrayList;
 import java.util.Arrays;
