+++ src/test/java/racingcar/CarListTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarList;
 
 import java.util.ArrayList;
 import java.util.List;
